+++ DuneGame/core/src/com/dune/game/core/Tank.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.g2d.*;
 import com.badlogic.gdx.math.*;
 import com.dune.game.core.Weapon.Type;
 
-public class Tank extends GameObject implements Poolable, Updatable {
+public class Tank extends GameObject implements Poolable {
   public enum Owner { PLAYER, AI }
   
   private Owner ownerType;
@@ -40,8 +40,7 @@ public class Tank extends GameObject implements Poolable, Updatable {
     this.weapon = new Weapon(Type.HARVEST, 3.0f, 1);
     this.destination = new Vector2(position);
   }
-  
-  @Override
+
   public void update(float dt) {
     if (Gdx.input.isButtonJustPressed(Input.Buttons.LEFT)
         && position.dst(Gdx.input.getX(), 720 - Gdx.input.getY()) < 40) {
@@ -103,8 +102,7 @@ public class Tank extends GameObject implements Poolable, Updatable {
     if (position.y < 40.0f) position.y = 40.0f;
     if (position.y > 680.0f) position.y = 680.0f;
   }
-  
-  @Override
+
   public void render(SpriteBatch batch) {
     batch.draw(textures[getCurrentFrameIndex()], position.x - 40, position.y - 40, 40, 40, 80, 80, 1, 1, angle);
     if (weapon.getType() == Weapon.Type.HARVEST && weapon.getUsageTimePercentage() > 0.0f) {
