+++ DuneGame/core/src/com/dune/game/core/TanksController.java
@@ -1,5 +1,7 @@
 package com.dune.game.core;
 
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+
 public class TanksController extends ObjectPool<Tank> {
   private GameController gc;
   
@@ -15,5 +17,18 @@ public class TanksController extends ObjectPool<Tank> {
   public void setup(float x, float y, Tank.Owner ownerType) {
     Tank t = getActiveElement();
     t.setup(x, y, ownerType);
+  }
+  
+  public void update(float dt) {
+    for (int i = 0; i < activeList.size(); i++) {
+      activeList.get(i).update(dt);
+    }
+    checkPool();
+  }
+  
+  public void render(SpriteBatch batch) {
+    for (int i = 0; i < activeList.size(); i++) {
+      activeList.get(i).render(batch);
+    }
   }
 }
